SHOP-105 <Consuming API endpoint to allow vendors to upload products - Part 1>
