+++ lib/screens/auth/login/login_screen.dart
@@ -76,9 +76,9 @@ class LoginScreen extends StatelessWidget {
                         SizedBox(
                           height: getPropHeight(16),
                         ),
-                        Row(
+                        const Row(
                           mainAxisAlignment: MainAxisAlignment.spaceAround,
-                          children: const [
+                          children: [
                             SocialCard(
                               icon: "assets/icons/facebook.svg",
                             ),
