+++ lib/screens/home/widgets/inventoryWidget.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:shoppa_app/constants/colors.dart';
 import 'package:shoppa_app/constants/constants.dart';
 import 'package:shoppa_app/constants/size_configurations.dart';
-import 'package:shoppa_app/screens/shop/addItemScreen.dart';
+import 'package:shoppa_app/screens/shop/uploadProductScreen.dart';
 import 'package:shoppa_app/widgets/inventoryItemCard.dart';
 import '../../../dummyData/all_dummy_data.dart';
 
@@ -77,7 +77,7 @@ class Inventory2 extends StatelessWidget {
                         borderRadius: BorderRadius.circular(8)),
                     color: primaryColor,
                     onPressed: () {
-                      Navigator.of(context).pushNamed(AddItemScreen.routeName);
+                      Navigator.of(context).pushNamed(UploadProductScreen.routeName);
                     },
                     child: Text(
                       " +  Add new Item ",
